+++ changelog.md
@@ -2,3 +2,4 @@
 ---
 
 - Fixed sign markers not being removed correctly when the sign is removed from the world.
+- Fixed feedback sounds not playing on Fabric.